+++ projects/ids-enterprise-ng/src/lib/calendar/index.ts
@@ -1,3 +1,2 @@
 export * from './soho-calendar.component';
-export * from './soho-calendar-monthview.component';
 export * from './soho-calendar.module';
